Fix potential error when updating the same index at the same time
Since there are 3+ repositories potentially trying to update the same index at the same time, it is possible that two GH Actions could try to create an index at the same time, causing a Meilisearch api error. This will in turn appear as if the continuous deployment was unsuccessful. This modification gracefully exits the update process that threw an exception, which is what we want since the other process will handle updating the index.